+++ Semester2/SZFA_LAB03/SZFA_LAB03/IRealEstate.cs
@@ -2,6 +2,6 @@
 {
     internal interface IRealEstate
     {
-        float TotalValue();
+        int TotalValue();
     }
 }
